requirements.txt
flask
flask-cors
sqlalchemy
pydantic
requests
faiss-cpu
sentence-transformers
groq  
streamlit
python-dotenv
datetime
collections
uuid
os
requests
json
re
openai